[New orden] Se borro archivos y solo se conservo la macro y el main
- Archivos Borrados: /Debug y archivos para usar el proyecto en Atmel
  Studio
