--- CAMPOS_BITS/CAMPOS_BITS/CAMPOS_BITS.componentinfo.xml
@@ -1,86 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Store xmlns:i="http://www.w3.org/2001/XMLSchema-instance" xmlns="AtmelPackComponentManagement">
-	<ProjectComponents>
-		<ProjectComponent z:Id="i1" xmlns:z="http://schemas.microsoft.com/2003/10/Serialization/">
-			<CApiVersion></CApiVersion>
-			<CBundle></CBundle>
-			<CClass>Device</CClass>
-			<CGroup>Startup</CGroup>
-			<CSub></CSub>
-			<CVariant></CVariant>
-			<CVendor>Atmel</CVendor>
-			<CVersion>1.3.0</CVersion>
-			<DefaultRepoPath>C:/Program Files (x86)\Atmel\Studio\7.0\Packs</DefaultRepoPath>
-			<DependentComponents xmlns:d4p1="http://schemas.microsoft.com/2003/10/Serialization/Arrays" />
-			<Description></Description>
-			<Files xmlns:d4p1="http://schemas.microsoft.com/2003/10/Serialization/Arrays">
-				<d4p1:anyType i:type="FileInfo">
-					<AbsolutePath>C:/Program Files (x86)\Atmel\Studio\7.0\Packs\atmel\ATmega_DFP\1.3.300\include</AbsolutePath>
-					<Attribute></Attribute>
-					<Category>include</Category>
-					<Condition>C</Condition>
-					<FileContentHash i:nil="true" />
-					<FileVersion></FileVersion>
-					<Name>include</Name>
-					<SelectString></SelectString>
-					<SourcePath></SourcePath>
-				</d4p1:anyType>
-				<d4p1:anyType i:type="FileInfo">
-					<AbsolutePath>C:/Program Files (x86)\Atmel\Studio\7.0\Packs\atmel\ATmega_DFP\1.3.300\include\avr\iom328p.h</AbsolutePath>
-					<Attribute></Attribute>
-					<Category>header</Category>
-					<Condition>C</Condition>
-					<FileContentHash>UMk4QUzkkuShabuoYtNl/Q==</FileContentHash>
-					<FileVersion></FileVersion>
-					<Name>include/avr/iom328p.h</Name>
-					<SelectString></SelectString>
-					<SourcePath></SourcePath>
-				</d4p1:anyType>
-				<d4p1:anyType i:type="FileInfo">
-					<AbsolutePath>C:/Program Files (x86)\Atmel\Studio\7.0\Packs\atmel\ATmega_DFP\1.3.300\templates\main.c</AbsolutePath>
-					<Attribute>template</Attribute>
-					<Category>source</Category>
-					<Condition>C Exe</Condition>
-					<FileContentHash>rrLy/3lJbql5p0HAtp18Rg==</FileContentHash>
-					<FileVersion></FileVersion>
-					<Name>templates/main.c</Name>
-					<SelectString>Main file (.c)</SelectString>
-					<SourcePath></SourcePath>
-				</d4p1:anyType>
-				<d4p1:anyType i:type="FileInfo">
-					<AbsolutePath>C:/Program Files (x86)\Atmel\Studio\7.0\Packs\atmel\ATmega_DFP\1.3.300\templates\main.cpp</AbsolutePath>
-					<Attribute>template</Attribute>
-					<Category>source</Category>
-					<Condition>C Exe</Condition>
-					<FileContentHash>YXFphlh0CtZJU+ebktABgQ==</FileContentHash>
-					<FileVersion></FileVersion>
-					<Name>templates/main.cpp</Name>
-					<SelectString>Main file (.cpp)</SelectString>
-					<SourcePath></SourcePath>
-				</d4p1:anyType>
-				<d4p1:anyType i:type="FileInfo">
-					<AbsolutePath>C:/Program Files (x86)\Atmel\Studio\7.0\Packs\atmel\ATmega_DFP\1.3.300\gcc\dev\atmega328p</AbsolutePath>
-					<Attribute></Attribute>
-					<Category>libraryPrefix</Category>
-					<Condition>GCC</Condition>
-					<FileContentHash i:nil="true" />
-					<FileVersion></FileVersion>
-					<Name>gcc/dev/atmega328p</Name>
-					<SelectString></SelectString>
-					<SourcePath></SourcePath>
-				</d4p1:anyType>
-			</Files>
-			<PackName>ATmega_DFP</PackName>
-			<PackPath>C:/Program Files (x86)/Atmel/Studio/7.0/Packs/atmel/ATmega_DFP/1.3.300/Atmel.ATmega_DFP.pdsc</PackPath>
-			<PackVersion>1.3.300</PackVersion>
-			<PresentInProject>true</PresentInProject>
-			<ReferenceConditionId>ATmega328P</ReferenceConditionId>
-			<RteComponents xmlns:d4p1="http://schemas.microsoft.com/2003/10/Serialization/Arrays">
-				<d4p1:string></d4p1:string>
-			</RteComponents>
-			<Status>Resolved</Status>
-			<VersionMode>Fixed</VersionMode>
-			<IsComponentInAtProject>true</IsComponentInAtProject>
-		</ProjectComponent>
-	</ProjectComponents>
-</Store>
