--- CAMPOS_BITS/CAMPOS_BITS/Debug/CAMPOS_BITS.eep
@@ -1,1 +1,0 @@
-:00000001FF
