--- CAMPOS_BITS/CAMPOS_BITS/Debug/CAMPOS_BITS.hex
@@ -1,15 +1,0 @@
-:100000000C9434000C943E000C943E000C943E0082
-:100010000C943E000C943E000C943E000C943E0068
-:100020000C943E000C943E000C943E000C943E0058
-:100030000C943E000C943E000C943E000C943E0048
-:100040000C943E000C943E000C943E000C943E0038
-:100050000C943E000C943E000C943E000C943E0028
-:100060000C943E000C943E0011241FBECFEFD8E04C
-:10007000DEBFCDBF0E9440000C9468000C940000CD
-:10008000389A259A539A40982D985B98409A2FEF6A
-:1000900083ED90E3215080409040E1F700C00000E4
-:1000A00040982D9A2FEF83ED90E3215080409040AF
-:1000B000E1F700C000002D985B9A2FEF83ED90E3ED
-:1000C000215080409040E1F700C000005B98DECFF7
-:0400D000F894FFCFD2
-:00000001FF
