--- CAMPOS_BITS/CAMPOS_BITS/Debug/CAMPOS_BITS.lss
@@ -1,151 +1,0 @@
-
-CAMPOS_BITS.elf:     file format elf32-avr
-
-Sections:
-Idx Name          Size      VMA       LMA       File off  Algn
-  0 .data         00000000  00800100  00800100  00000128  2**0
-                  CONTENTS, ALLOC, LOAD, DATA
-  1 .text         000000d4  00000000  00000000  00000054  2**1
-                  CONTENTS, ALLOC, LOAD, READONLY, CODE
-  2 .comment      00000030  00000000  00000000  00000128  2**0
-                  CONTENTS, READONLY
-  3 .note.gnu.avr.deviceinfo 00000040  00000000  00000000  00000158  2**2
-                  CONTENTS, READONLY
-  4 .debug_aranges 00000020  00000000  00000000  00000198  2**0
-                  CONTENTS, READONLY, DEBUGGING
-  5 .debug_info   00000b7c  00000000  00000000  000001b8  2**0
-                  CONTENTS, READONLY, DEBUGGING
-  6 .debug_abbrev 000006a3  00000000  00000000  00000d34  2**0
-                  CONTENTS, READONLY, DEBUGGING
-  7 .debug_line   000002d8  00000000  00000000  000013d7  2**0
-                  CONTENTS, READONLY, DEBUGGING
-  8 .debug_frame  00000024  00000000  00000000  000016b0  2**2
-                  CONTENTS, READONLY, DEBUGGING
-  9 .debug_str    000004fe  00000000  00000000  000016d4  2**0
-                  CONTENTS, READONLY, DEBUGGING
- 10 .debug_loc    000000d8  00000000  00000000  00001bd2  2**0
-                  CONTENTS, READONLY, DEBUGGING
- 11 .debug_ranges 00000010  00000000  00000000  00001caa  2**0
-                  CONTENTS, READONLY, DEBUGGING
-
-Disassembly of section .text:
-
-00000000 <__vectors>:
-   0:	0c 94 34 00 	jmp	0x68	; 0x68 <__ctors_end>
-   4:	0c 94 3e 00 	jmp	0x7c	; 0x7c <__bad_interrupt>
-   8:	0c 94 3e 00 	jmp	0x7c	; 0x7c <__bad_interrupt>
-   c:	0c 94 3e 00 	jmp	0x7c	; 0x7c <__bad_interrupt>
-  10:	0c 94 3e 00 	jmp	0x7c	; 0x7c <__bad_interrupt>
-  14:	0c 94 3e 00 	jmp	0x7c	; 0x7c <__bad_interrupt>
-  18:	0c 94 3e 00 	jmp	0x7c	; 0x7c <__bad_interrupt>
-  1c:	0c 94 3e 00 	jmp	0x7c	; 0x7c <__bad_interrupt>
-  20:	0c 94 3e 00 	jmp	0x7c	; 0x7c <__bad_interrupt>
-  24:	0c 94 3e 00 	jmp	0x7c	; 0x7c <__bad_interrupt>
-  28:	0c 94 3e 00 	jmp	0x7c	; 0x7c <__bad_interrupt>
-  2c:	0c 94 3e 00 	jmp	0x7c	; 0x7c <__bad_interrupt>
-  30:	0c 94 3e 00 	jmp	0x7c	; 0x7c <__bad_interrupt>
-  34:	0c 94 3e 00 	jmp	0x7c	; 0x7c <__bad_interrupt>
-  38:	0c 94 3e 00 	jmp	0x7c	; 0x7c <__bad_interrupt>
-  3c:	0c 94 3e 00 	jmp	0x7c	; 0x7c <__bad_interrupt>
-  40:	0c 94 3e 00 	jmp	0x7c	; 0x7c <__bad_interrupt>
-  44:	0c 94 3e 00 	jmp	0x7c	; 0x7c <__bad_interrupt>
-  48:	0c 94 3e 00 	jmp	0x7c	; 0x7c <__bad_interrupt>
-  4c:	0c 94 3e 00 	jmp	0x7c	; 0x7c <__bad_interrupt>
-  50:	0c 94 3e 00 	jmp	0x7c	; 0x7c <__bad_interrupt>
-  54:	0c 94 3e 00 	jmp	0x7c	; 0x7c <__bad_interrupt>
-  58:	0c 94 3e 00 	jmp	0x7c	; 0x7c <__bad_interrupt>
-  5c:	0c 94 3e 00 	jmp	0x7c	; 0x7c <__bad_interrupt>
-  60:	0c 94 3e 00 	jmp	0x7c	; 0x7c <__bad_interrupt>
-  64:	0c 94 3e 00 	jmp	0x7c	; 0x7c <__bad_interrupt>
-
-00000068 <__ctors_end>:
-  68:	11 24       	eor	r1, r1
-  6a:	1f be       	out	0x3f, r1	; 63
-  6c:	cf ef       	ldi	r28, 0xFF	; 255
-  6e:	d8 e0       	ldi	r29, 0x08	; 8
-  70:	de bf       	out	0x3e, r29	; 62
-  72:	cd bf       	out	0x3d, r28	; 61
-  74:	0e 94 40 00 	call	0x80	; 0x80 <main>
-  78:	0c 94 68 00 	jmp	0xd0	; 0xd0 <_exit>
-
-0000007c <__bad_interrupt>:
-  7c:	0c 94 00 00 	jmp	0	; 0x0 <__vectors>
-
-00000080 <main>:
-#include <util/delay.h>
-#include "DEF_ATMEGA328P.h"
-
-int main(void)
-{
-	DDRCbits.DDRC0 = 1;
-  80:	38 9a       	sbi	0x07, 0	; 7
-	DDRBbits.DDRB5 = 1;
-  82:	25 9a       	sbi	0x04, 5	; 4
-	DDRDbits.DDRD3 = 1;
-  84:	53 9a       	sbi	0x0a, 3	; 10
-	
-	PORTCbits.PC0_ = 0;
-  86:	40 98       	cbi	0x08, 0	; 8
-	PORTBbits.PB5_ = 0;		//Inicia en bajo
-  88:	2d 98       	cbi	0x05, 5	; 5
-	PORTDbits.PD3_ = 0;
-  8a:	5b 98       	cbi	0x0b, 3	; 11
-	
-    /* Replace with your application code */
-    while (1) 
-    {
-		PORTCbits.PC0_ = 1;
-  8c:	40 9a       	sbi	0x08, 0	; 8
-	#else
-		//round up by default
-		__ticks_dc = (uint32_t)(ceil(fabs(__tmp)));
-	#endif
-
-	__builtin_avr_delay_cycles(__ticks_dc);
-  8e:	2f ef       	ldi	r18, 0xFF	; 255
-  90:	83 ed       	ldi	r24, 0xD3	; 211
-  92:	90 e3       	ldi	r25, 0x30	; 48
-  94:	21 50       	subi	r18, 0x01	; 1
-  96:	80 40       	sbci	r24, 0x00	; 0
-  98:	90 40       	sbci	r25, 0x00	; 0
-  9a:	e1 f7       	brne	.-8      	; 0x94 <main+0x14>
-  9c:	00 c0       	rjmp	.+0      	; 0x9e <main+0x1e>
-  9e:	00 00       	nop
-		_delay_ms(1000);
-		PORTCbits.PC0_ = 0;
-  a0:	40 98       	cbi	0x08, 0	; 8
-		PORTBbits.PB5_ = 1;
-  a2:	2d 9a       	sbi	0x05, 5	; 5
-  a4:	2f ef       	ldi	r18, 0xFF	; 255
-  a6:	83 ed       	ldi	r24, 0xD3	; 211
-  a8:	90 e3       	ldi	r25, 0x30	; 48
-  aa:	21 50       	subi	r18, 0x01	; 1
-  ac:	80 40       	sbci	r24, 0x00	; 0
-  ae:	90 40       	sbci	r25, 0x00	; 0
-  b0:	e1 f7       	brne	.-8      	; 0xaa <main+0x2a>
-  b2:	00 c0       	rjmp	.+0      	; 0xb4 <main+0x34>
-  b4:	00 00       	nop
-		_delay_ms(1000);
-		PORTBbits.PB5_ = 0;
-  b6:	2d 98       	cbi	0x05, 5	; 5
-		PORTDbits.PD3_ = 1;
-  b8:	5b 9a       	sbi	0x0b, 3	; 11
-  ba:	2f ef       	ldi	r18, 0xFF	; 255
-  bc:	83 ed       	ldi	r24, 0xD3	; 211
-  be:	90 e3       	ldi	r25, 0x30	; 48
-  c0:	21 50       	subi	r18, 0x01	; 1
-  c2:	80 40       	sbci	r24, 0x00	; 0
-  c4:	90 40       	sbci	r25, 0x00	; 0
-  c6:	e1 f7       	brne	.-8      	; 0xc0 <main+0x40>
-  c8:	00 c0       	rjmp	.+0      	; 0xca <main+0x4a>
-  ca:	00 00       	nop
-		_delay_ms(1000);
-		PORTDbits.PD3_ = 0;	
-  cc:	5b 98       	cbi	0x0b, 3	; 11
-  ce:	de cf       	rjmp	.-68     	; 0x8c <main+0xc>
-
-000000d0 <_exit>:
-  d0:	f8 94       	cli
-
-000000d2 <__stop_program>:
-  d2:	ff cf       	rjmp	.-2      	; 0xd2 <__stop_program>
