--- CAMPOS_BITS/CAMPOS_BITS/Debug/CAMPOS_BITS.srec
@@ -1,16 +1,0 @@
-S013000043414D504F535F424954532E73726563BD
-S11300000C9434000C943E000C943E000C943E007E
-S11300100C943E000C943E000C943E000C943E0064
-S11300200C943E000C943E000C943E000C943E0054
-S11300300C943E000C943E000C943E000C943E0044
-S11300400C943E000C943E000C943E000C943E0034
-S11300500C943E000C943E000C943E000C943E0024
-S11300600C943E000C943E0011241FBECFEFD8E048
-S1130070DEBFCDBF0E9440000C9468000C940000C9
-S1130080389A259A539A40982D985B98409A2FEF66
-S113009083ED90E3215080409040E1F700C00000E0
-S11300A040982D9A2FEF83ED90E3215080409040AB
-S11300B0E1F700C000002D985B9A2FEF83ED90E3E9
-S11300C0215080409040E1F700C000005B98DECFF3
-S10700D0F894FFCFCE
-S9030000FC
