--- CAMPOS_BITS/CAMPOS_BITS/Debug/Makefile
@@ -1,127 +1,0 @@
-################################################################################
-# Automatically-generated file. Do not edit!
-################################################################################
-
-SHELL := cmd.exe
-RM := rm -rf
-
-USER_OBJS :=
-
-LIBS := 
-PROJ := 
-
-O_SRCS := 
-C_SRCS := 
-S_SRCS := 
-S_UPPER_SRCS := 
-OBJ_SRCS := 
-ASM_SRCS := 
-PREPROCESSING_SRCS := 
-OBJS := 
-OBJS_AS_ARGS := 
-C_DEPS := 
-C_DEPS_AS_ARGS := 
-EXECUTABLES := 
-OUTPUT_FILE_PATH :=
-OUTPUT_FILE_PATH_AS_ARGS :=
-AVR_APP_PATH :=$$$AVR_APP_PATH$$$
-QUOTE := "
-ADDITIONAL_DEPENDENCIES:=
-OUTPUT_FILE_DEP:=
-LIB_DEP:=
-LINKER_SCRIPT_DEP:=
-
-# Every subdirectory with source files must be described here
-SUBDIRS := 
-
-
-# Add inputs and outputs from these tool invocations to the build variables 
-C_SRCS +=  \
-../main.c
-
-
-PREPROCESSING_SRCS += 
-
-
-ASM_SRCS += 
-
-
-OBJS +=  \
-main.o
-
-OBJS_AS_ARGS +=  \
-main.o
-
-C_DEPS +=  \
-main.d
-
-C_DEPS_AS_ARGS +=  \
-main.d
-
-OUTPUT_FILE_PATH +=CAMPOS_BITS.elf
-
-OUTPUT_FILE_PATH_AS_ARGS +=CAMPOS_BITS.elf
-
-ADDITIONAL_DEPENDENCIES:=
-
-OUTPUT_FILE_DEP:= ./makedep.mk
-
-LIB_DEP+= 
-
-LINKER_SCRIPT_DEP+= 
-
-
-# AVR32/GNU C Compiler
-./main.o: .././main.c
-	@echo Building file: $<
-	@echo Invoking: AVR/GNU C Compiler : 5.4.0
-	$(QUOTE)C:\Program Files (x86)\Atmel\Studio\7.0\toolchain\avr8\avr8-gnu-toolchain\bin\avr-gcc.exe$(QUOTE)  -x c -funsigned-char -funsigned-bitfields -DDEBUG  -I"C:\Program Files (x86)\Atmel\Studio\7.0\Packs\atmel\ATmega_DFP\1.3.300\include"  -O1 -ffunction-sections -fdata-sections -fpack-struct -fshort-enums -g2 -Wall -mmcu=atmega328p -B "C:\Program Files (x86)\Atmel\Studio\7.0\Packs\atmel\ATmega_DFP\1.3.300\gcc\dev\atmega328p" -c -std=gnu99 -MD -MP -MF "$(@:%.o=%.d)" -MT"$(@:%.o=%.d)" -MT"$(@:%.o=%.o)"   -o "$@" "$<" 
-	@echo Finished building: $<
-	
-
-
-
-
-# AVR32/GNU Preprocessing Assembler
-
-
-
-# AVR32/GNU Assembler
-
-
-
-
-ifneq ($(MAKECMDGOALS),clean)
-ifneq ($(strip $(C_DEPS)),)
--include $(C_DEPS)
-endif
-endif
-
-# Add inputs and outputs from these tool invocations to the build variables 
-
-# All Target
-all: $(OUTPUT_FILE_PATH) $(ADDITIONAL_DEPENDENCIES)
-
-$(OUTPUT_FILE_PATH): $(OBJS) $(USER_OBJS) $(OUTPUT_FILE_DEP) $(LIB_DEP) $(LINKER_SCRIPT_DEP)
-	@echo Building target: $@
-	@echo Invoking: AVR/GNU Linker : 5.4.0
-	$(QUOTE)C:\Program Files (x86)\Atmel\Studio\7.0\toolchain\avr8\avr8-gnu-toolchain\bin\avr-gcc.exe$(QUOTE) -o$(OUTPUT_FILE_PATH_AS_ARGS) $(OBJS_AS_ARGS) $(USER_OBJS) $(LIBS) -Wl,-Map="CAMPOS_BITS.map" -Wl,--start-group -Wl,-lm  -Wl,--end-group -Wl,--gc-sections -mmcu=atmega328p -B "C:\Program Files (x86)\Atmel\Studio\7.0\Packs\atmel\ATmega_DFP\1.3.300\gcc\dev\atmega328p"  
-	@echo Finished building target: $@
-	"C:\Program Files (x86)\Atmel\Studio\7.0\toolchain\avr8\avr8-gnu-toolchain\bin\avr-objcopy.exe" -O ihex -R .eeprom -R .fuse -R .lock -R .signature -R .user_signatures  "CAMPOS_BITS.elf" "CAMPOS_BITS.hex"
-	"C:\Program Files (x86)\Atmel\Studio\7.0\toolchain\avr8\avr8-gnu-toolchain\bin\avr-objcopy.exe" -j .eeprom  --set-section-flags=.eeprom=alloc,load --change-section-lma .eeprom=0  --no-change-warnings -O ihex "CAMPOS_BITS.elf" "CAMPOS_BITS.eep" || exit 0
-	"C:\Program Files (x86)\Atmel\Studio\7.0\toolchain\avr8\avr8-gnu-toolchain\bin\avr-objdump.exe" -h -S "CAMPOS_BITS.elf" > "CAMPOS_BITS.lss"
-	"C:\Program Files (x86)\Atmel\Studio\7.0\toolchain\avr8\avr8-gnu-toolchain\bin\avr-objcopy.exe" -O srec -R .eeprom -R .fuse -R .lock -R .signature -R .user_signatures "CAMPOS_BITS.elf" "CAMPOS_BITS.srec"
-	"C:\Program Files (x86)\Atmel\Studio\7.0\toolchain\avr8\avr8-gnu-toolchain\bin\avr-size.exe" "CAMPOS_BITS.elf"
-	
-	
-
-
-
-
-
-# Other Targets
-clean:
-	-$(RM) $(OBJS_AS_ARGS) $(EXECUTABLES)  
-	-$(RM) $(C_DEPS_AS_ARGS)   
-	rm -rf "CAMPOS_BITS.elf" "CAMPOS_BITS.a" "CAMPOS_BITS.hex" "CAMPOS_BITS.lss" "CAMPOS_BITS.eep" "CAMPOS_BITS.map" "CAMPOS_BITS.srec" "CAMPOS_BITS.usersignatures"
-
